backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
supabase==2.3.4
python-docx==1.1.0
requests==2.31.0
python-jose==3.3.0 